[NCL-8539] rename package
