+++ bpm/src/test/java/org/jboss/pnc/bpm/BuildResultPushManagerTest.java
@@ -23,9 +23,9 @@ import org.jboss.pnc.bpm.causeway.InProgress;
 import org.jboss.pnc.bpm.causeway.Result;
 import org.jboss.pnc.causewayclient.CausewayClient;
 import org.jboss.pnc.common.concurrent.Sequence;
-import org.jboss.pnc.common.gerrit.ScmUrlGeneratorProvider;
-import org.jboss.pnc.common.gerrit.ScmException;
-import org.jboss.pnc.common.gerrit.ScmUrlGenerator;
+import org.jboss.pnc.common.scm.ScmUrlGeneratorProvider;
+import org.jboss.pnc.common.scm.ScmException;
+import org.jboss.pnc.common.scm.ScmUrlGenerator;
 import org.jboss.pnc.dto.BuildPushResult;
 import org.jboss.pnc.enums.BuildPushStatus;
 import org.jboss.pnc.enums.BuildStatus;
@@ -52,7 +52,7 @@ import java.util.Collections;
 import java.util.Date;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.jboss.pnc.common.gerrit.ScmUrlGeneratorProvider.SCMProvider.GERRIT;
+import static org.jboss.pnc.common.scm.ScmUrlGeneratorProvider.SCMProvider.GERRIT;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
