+++ common/src/main/java/org/jboss/pnc/common/scm/GitlabScmUrlGenerator.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.common.gerrit;
+package org.jboss.pnc.common.scm;
 
 import javax.validation.constraints.NotNull;
 import java.net.URI;
