+++ facade/src/main/java/org/jboss/pnc/facade/providers/BuildProviderImpl.java
@@ -75,8 +75,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.Getter;
 import org.jboss.pnc.auth.KeycloakServiceClient;
-import org.jboss.pnc.common.gerrit.ScmUrlGeneratorProvider;
-import org.jboss.pnc.common.gerrit.ScmException;
+import org.jboss.pnc.common.scm.ScmUrlGeneratorProvider;
+import org.jboss.pnc.common.scm.ScmException;
 import org.jboss.pnc.common.graph.GraphBuilder;
 import org.jboss.pnc.common.graph.GraphUtils;
 import org.jboss.pnc.common.pnc.LongBase32IdConverter;
